project/py/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "olik-font"
version = "0.1.0"
description = "Glyph scene graph — pass-1 implementation"
requires-python = ">=3.11"
dependencies = [
  "pyyaml>=6.0",
  "jsonschema>=4.21",
  "requests>=2.31",
  "svgpathtools>=1.6",
  "shapely>=2.0",
  "python-dateutil>=2.9",  # required by quicktype-generated types
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.6",
]

[project.scripts]
olik = "olik_font.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/olik_font"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-ra --strict-markers"

[tool.ruff]
target-version = "py311"
line-length = 100
# Generated code is regenerated via `task codegen:*` — don't lint/format it.
extend-exclude = [".venv", "data", "src/olik_font/generated"]

[tool.ruff.lint]
# Curated rule set: pyflakes, pycodestyle, import sorting, naming, warnings,
# pyupgrade, bugbear, comprehensions, simplify, ruff-specific. Skip docstring
# rules and strict typing rules to keep iteration friction low.
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM", "RUF"]
ignore = [
  "E501",  # line length — enforced by formatter, not linter
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
  "N806",  # non-lowercase variable names fine in tests (e.g., for fixtures)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
